agg/requirements.txt
cassandra-driver==3.25.0
click==8.1.3
geomet==0.2.1.post1
pymongo==4.1.1
six==1.16.0
